Adaption to work well with ReflactionBasedAbstractFactory
New initializers for Navigation and ViewRenderer
Controller plugin for Navigation
Add the initializers to the AbstractController
Abstract table now expects TableGatewayInterface so that it kan easily be modified in descending classes.
AbstractTableGateway to simplify auto wiring scheme.
Don't use current user in LogTable as this package has no clear definition of users.